Added Modules and main class
